Restructured project directories
